Ensure MSI generation does not silently fail V4 (#2967)
* Ensure MSI generation does not silently fail and arguments are correctly parsed

* Check the .msi path

* Removed debugging statements

+++ generateMsiFiles.ps1
@@ -42,8 +42,14 @@ $cliVersion = [System.Diagnostics.FileVersionInfo]::GetVersionInfo($cli).FileVer
     $msiPath = "$artifactsPath\$msiName.msi"
 
     & { heat dir '.' -cg FuncHost -dr INSTALLDIR -gg -ke -out $fragmentPath -srd -sreg -template fragment -var var.Source }
-    & { candle -arch $platform -dPlatform='$platform' -dSource='.' -dProductVersion='$cliVersion' $masterWxsPath $fragmentPath }
-    & { light -ext WixUIExtension -out $msiPath -sice:ICE61 $masterWxsName.wixobj $fragmentName.wixobj }
+    & { candle -arch $platform -dPlatform="$platform" -dSource='.' -dProductVersion="$cliVersion" $masterWxsPath $fragmentPath }
+    & { light -ext "WixUIExtension" -out $msiPath -sice:"ICE61" "$masterWxsName.wixobj" "$fragmentName.wixobj" }
+    
+    # Check that the .msi files are actually present
+    if (-not(Test-Path -Path $msiPath))
+    {
+        throw "$msiPath not found."
+    }
 
     Set-Location $baseDir
     Get-ChildItem -Path $targetDir -Recurse | Remove-Item -Force -Recurse -ea SilentlyContinue
